requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
idna==2.10
    # via requests
python-dateutil==2.8.1
    # via vobject
requests==2.25.1
    # via wekanapi
six==1.15.0
    # via python-dateutil
urllib3==1.26.3
    # via requests
vobject==0.9.6.1
    # via -r requirements.in
git+https://github.com/wekan/wekan-python-api-client.git@82ae0b685fc2bd74a52401e2716551e59a3df019#subdirectory=src
    # via -r requirements.in
flask
